pyproject.toml
[project]
name = "qmdr"
version = "0.1.0"
description = "A Flet desktop GUI for QQ Music downloads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "flet[all]>=0.85.0",
    "mutagen>=1.47.0",
    "qqmusic-api-python==0.3.6",
]
